% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unix_time.R
\name{convert_unix_time}
\alias{convert_unix_time}
\title{Convert UNIX time}
\usage{
convert_unix_time(x)
}
\arguments{
\item{x}{number or string}
}
\description{
Convert date-time from milliseconds UTC
}
