% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_features.R
\name{get_artist_features}
\alias{get_artist_features}
\title{Get artist features}
\usage{
get_artist_features(ids)
}
\arguments{
\item{ids}{character vector, Spotify artist IDs}

\item{as.df}{boolean, return data.frame of results or parsed list
(as-is from the API call)}
}
\description{
Get artist features
}
