% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recently_played.R
\name{get_recently_played}
\alias{get_recently_played}
\title{Get recently played tracks}
\usage{
get_recently_played(limit = 50, before = NULL, after = NULL)
}
\arguments{
\item{limit}{numeric, limit the number of records returned (max = 50)}

\item{before}{numeric, recently played tracks before this time (in milliseconds UTC)}

\item{after}{numeric, recently played tracks after this time (in milliseconds UTC)}
}
\value{
parsed list from the GET
}
\description{
Get recently played tracks
}
