% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_track_card.R
\name{make_track_card}
\alias{make_track_card}
\title{Make a track information card - HTML object}
\usage{
make_track_card(summarised_track)
}
\arguments{
\item{summarised_track}{data.frame, a row from summarise_recently_played}
}
\description{
HTML
}
\examples{
\dontrun{
get_recently_played() \%>\%
  summarise_recently_played() \%>\%
  slice(1) \%>\%
  make_track_card() \%>\%
  htmltools::browsable()}
}
