% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_recently_played.R
\name{summarise_recently_played}
\alias{summarise_recently_played}
\title{Summarise (wrangle) a recently played API call result}
\usage{
summarise_recently_played(result_list)
}
\arguments{
\item{result_list}{list, that from get_recently_played}
}
\value{
data.frame
}
\description{
The API call result is a list of data and this function turns parts of it into
a data.frame for ease of use
}
